requirements.txt
asgiref==3.3.1
astroid==2.5.6
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.1.0
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2-binary==2.9.1
pylint==2.8.3
pytz==2020.5
rope==0.18.0
selenium==3.141.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
